# openapi.json
{"openapi": "3.1.0", "info": {"title": "Hallal Invest API", "version": "0.1.0"}, "paths": {"/health": {"get": {"tags": ["meta"], "summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Health Get"}}}}}}}, "/api/v1/ping": {"get": {"tags": ["meta"], "summary": "Ping", "operationId": "ping_api_v1_ping_get", "parameters": [{"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"type": "boolean"}, "title": "Response Ping Api V1 Ping Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/shariah/{symbol}": {"get": {"tags": ["shariah"], "summary": "Shariah Screen", "operationId": "shariah_screen_api_v1_shariah__symbol__get", "parameters": [{"name": "symbol", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 20, "title": "Symbol"}}, {"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShariahReport"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/financials/{ticker}": {"get": {"tags": ["financials"], "summary": "Financials Snapshot", "operationId": "financials_snapshot_api_v1_financials__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 20, "title": "Ticker"}}, {"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FinancialsSnapshot"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/transactions": {"post": {"tags": ["portfolio"], "summary": "Create Transaction", "operationId": "create_transaction_api_v1_portfolio_transactions_post", "parameters": [{"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["portfolio"], "summary": "List Transactions", "operationId": "list_transactions_api_v1_portfolio_transactions_get", "parameters": [{"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TransactionRead"}, "title": "Response List Transactions Api V1 Portfolio Transactions Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/transactions/{tx_id}": {"delete": {"tags": ["portfolio"], "summary": "Delete Transaction", "operationId": "delete_transaction_api_v1_portfolio_transactions__tx_id__delete", "parameters": [{"name": "tx_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Tx Id"}}, {"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/positions": {"get": {"tags": ["portfolio"], "summary": "List Positions", "operationId": "list_positions_api_v1_portfolio_positions_get", "parameters": [{"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/PositionRead"}, "title": "Response List Positions Api V1 Portfolio Positions Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/summary": {"get": {"tags": ["portfolio"], "summary": "Get Summary", "operationId": "get_summary_api_v1_portfolio_summary_get", "parameters": [{"name": "x-auth-token", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X-Auth-Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PortfolioSummary"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"FinancialsSnapshot": {"properties": {"ticker": {"type": "string", "title": "Ticker", "description": "Ticker normalised to uppercase"}, "cik": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Cik", "description": "SEC Central Index Key, None if NOT_COVERED"}, "entity_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Entity Name"}, "verdict": {"type": "string", "enum": ["AVAILABLE", "NOT_COVERED", "ERROR"], "title": "Verdict"}, "fiscal_year": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Fiscal Year"}, "period_end": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Period End"}, "form": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Form", "description": "e.g. \"10-K\""}, "accession_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Accession Number"}, "filed": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Filed"}, "revenues": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Revenues"}, "net_income": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Net Income"}, "total_assets": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Total Assets"}, "long_term_debt": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Long Term Debt"}, "short_term_borrowings": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Short Term Borrowings"}, "cash_and_equivalents": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Cash And Equivalents"}, "interest_income_operating": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Interest Income Operating"}, "eps_diluted": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Eps Diluted"}, "operating_cash_flow": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Operating Cash Flow"}, "capex": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Capex"}, "stockholders_equity": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Stockholders Equity"}, "source": {"type": "string", "enum": ["SEC EDGAR API", "cache (financials_cache)"], "title": "Source"}, "cached": {"type": "boolean", "title": "Cached", "default": false}, "cache_age_hours": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Cache Age Hours"}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}}, "type": "object", "required": ["ticker", "verdict", "source"], "title": "FinancialsSnapshot", "description": "Latest annual snapshot extracted from SEC EDGAR for a US-listed ticker.\n\nVerdict semantics:\n  - AVAILABLE   : metadata + 11 concept values populated (some may be\n                  None if the concept is not reported by the company)\n  - NOT_COVERED : ticker not in SEC EDGAR universe (non-US, unknown,\n                  or de-listed)\n  - ERROR       : transient upstream failure (timeout / 5xx / network)"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MethodologyVerdict": {"properties": {"status": {"type": "string", "enum": ["compliant", "non_compliant"], "title": "Status"}, "failed_ratios": {"items": {"type": "string"}, "type": "array", "title": "Failed Ratios"}}, "type": "object", "required": ["status"], "title": "MethodologyVerdict", "description": "Verdict from one of the 5 published methodologies (\u00a73.2.1)."}, "PortfolioSummary": {"properties": {"total_invested": {"type": "string", "title": "Total Invested"}, "current_value": {"type": "string", "title": "Current Value"}, "unrealized_pnl": {"type": "string", "title": "Unrealized Pnl"}, "realized_pnl": {"type": "string", "title": "Realized Pnl"}, "total_pnl": {"type": "string", "title": "Total Pnl"}, "pnl_pct": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pnl Pct", "description": "(unrealized + realized) / total_invested. None if invested == 0."}, "n_positions": {"type": "integer", "title": "N Positions", "description": "Open positions with quantity > 0."}, "n_transactions": {"type": "integer", "title": "N Transactions"}, "currency": {"type": "string", "title": "Currency", "default": "USD"}}, "type": "object", "required": ["total_invested", "current_value", "unrealized_pnl", "realized_pnl", "total_pnl", "n_positions", "n_transactions"], "title": "PortfolioSummary", "description": "Aggregate snapshot across all positions (USD assumption, V1)."}, "PositionRead": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "symbol": {"type": "string", "title": "Symbol"}, "currency": {"type": "string", "title": "Currency"}, "opened_at": {"type": "string", "format": "date-time", "title": "Opened At"}, "quantity": {"type": "string", "title": "Quantity"}, "avg_cost": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Avg Cost", "description": "Weighted-average cost of currently-held shares (fees included). None when quantity == 0."}, "cost_basis": {"type": "string", "title": "Cost Basis", "description": "Total cost of currently-held shares = quantity \u00d7 avg_cost."}, "last_price": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Price", "description": "PROXY for the current market price: the most recent transaction price on this position. NOT a real-time quote. Replaced by a YFinance/equivalent feed in \u00e9tape 5."}, "current_value": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Value", "description": "quantity \u00d7 last_price (None if no transactions or qty==0)."}, "unrealized_pnl": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Unrealized Pnl", "description": "current_value \u2212 cost_basis (None if current_value is None)."}, "realized_pnl": {"type": "string", "title": "Realized Pnl", "description": "Cumulated profit/loss from SELL transactions on this position."}, "transactions_count": {"type": "integer", "title": "Transactions Count"}}, "type": "object", "required": ["id", "symbol", "currency", "opened_at", "quantity", "cost_basis", "realized_pnl", "transactions_count"], "title": "PositionRead", "description": "A position with all derived fields computed on the fly.\n\nAll Decimal fields are USD (single-currency in V1). When the position\nis closed (``quantity == 0``), the average-cost-dependent fields\n(``avg_cost``, ``current_value``, ``unrealized_pnl``) become ``None``\nwhile ``realized_pnl`` keeps the accumulated P&L over the lifetime\nof the position."}, "ShariahCheck": {"properties": {"value": {"type": "number", "title": "Value"}, "threshold": {"type": "number", "title": "Threshold"}, "pass": {"type": "boolean", "title": "Pass", "description": "True if value <= threshold (alias `pass` matches \u00a75.2)"}}, "type": "object", "required": ["value", "threshold", "pass"], "title": "ShariahCheck", "description": "One of the 4 bloquant checks evaluated against \u00a75.1 thresholds.\n\nPopulated only when the upstream surfaces raw ratios. Empty in\nfree-tier mode \u2014 see schema docstring."}, "ShariahRatios": {"properties": {"debt_to_marketcap": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Debt To Marketcap"}, "debt_to_assets": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Debt To Assets"}, "cash_to_marketcap": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Cash To Marketcap"}, "impure_revenue_ratio": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Impure Revenue Ratio"}, "interest_income_ratio": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Interest Income Ratio"}, "receivables_to_assets": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Receivables To Assets"}, "non_compliant_assets_ratio": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Non Compliant Assets Ratio"}}, "additionalProperties": true, "type": "object", "title": "ShariahRatios", "description": "Raw ratios returned by Halal Terminal \u00a73.2.1.\n\nAll fields optional: providers may omit any subset depending on\ncoverage. The free tier omits all of them \u2014 see schema docstring."}, "ShariahReport": {"properties": {"symbol": {"type": "string", "title": "Symbol", "description": "Ticker normalised to uppercase"}, "verdict": {"type": "string", "enum": ["PASS", "FAIL", "ERROR", "NOT_COVERED"], "title": "Verdict"}, "failed_checks": {"items": {"type": "string", "enum": ["debt_to_marketcap", "cash_to_marketcap", "impure_revenue_ratio", "interest_income_ratio"]}, "type": "array", "title": "Failed Checks"}, "checks": {"additionalProperties": {"$ref": "#/components/schemas/ShariahCheck"}, "propertyNames": {"enum": ["debt_to_marketcap", "cash_to_marketcap", "impure_revenue_ratio", "interest_income_ratio"]}, "type": "object", "title": "Checks"}, "halal_terminal_methodology_verdicts": {"additionalProperties": {"$ref": "#/components/schemas/MethodologyVerdict"}, "type": "object", "title": "Halal Terminal Methodology Verdicts"}, "raw_ratios": {"$ref": "#/components/schemas/ShariahRatios"}, "as_of_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "As Of Date"}, "source": {"type": "string", "enum": ["Halal Terminal API (aggregate verdict)", "Halal Terminal API (error)", "cache (screen_history)"], "title": "Source"}, "cached": {"type": "boolean", "title": "Cached", "default": false}, "cache_age_days": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Cache Age Days"}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}}, "type": "object", "required": ["symbol", "verdict", "source"], "title": "ShariahReport", "description": "Response of `GET /api/v1/shariah/{symbol}`.\n\nMirrors \u00a75.2 output, extended with cache metadata (see module\ndocstring). In free-tier mode, the 4-checks / methodology /\nraw_ratios fields are empty by design."}, "TransactionCreate": {"properties": {"symbol": {"type": "string", "maxLength": 20, "minLength": 1, "title": "Symbol"}, "currency": {"type": "string", "maxLength": 3, "minLength": 3, "title": "Currency", "default": "USD"}, "date": {"type": "string", "format": "date", "title": "Date"}, "qty": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Qty"}, "price": {"anyOf": [{"type": "number", "exclusiveMinimum": 0.0}, {"type": "string"}], "title": "Price"}, "fee": {"anyOf": [{"type": "number", "minimum": 0.0}, {"type": "string"}], "title": "Fee", "default": "0"}}, "additionalProperties": false, "type": "object", "required": ["symbol", "date", "qty", "price"], "title": "TransactionCreate", "description": "Body of POST /api/v1/portfolio/transactions.\n\nValidation:\n  - symbol non-empty, \u2264 20 chars, auto-uppercased.\n  - currency 3 chars, auto-uppercased. Service rejects \u2260 \"USD\" in V1.\n  - date is the trade date (accept future dates for V1 \u2014 scheduled-buy\n    use-case, no validation).\n  - qty signed, non-zero (positive=BUY, negative=SELL).\n  - price strictly positive.\n  - fee non-negative."}, "TransactionRead": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "position_id": {"type": "string", "format": "uuid", "title": "Position Id"}, "symbol": {"type": "string", "title": "Symbol"}, "currency": {"type": "string", "title": "Currency"}, "date": {"type": "string", "format": "date", "title": "Date"}, "qty": {"type": "string", "title": "Qty"}, "price": {"type": "string", "title": "Price"}, "fee": {"type": "string", "title": "Fee"}}, "type": "object", "required": ["id", "position_id", "symbol", "currency", "date", "qty", "price", "fee"], "title": "TransactionRead", "description": "A persisted transaction row plus joined position metadata."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}